Refactor ShutdownHandler to new module
